+++ fw-cloud-business/fw-cloud-business-admin/src/main/java/com/github/liuweijw/business/admin/service/impl/DeptServiceImpl.java
@@ -42,4 +42,5 @@ public class DeptServiceImpl extends JPAFactoryImpl implements DeptService {
 		}
 		return TreeUtil.build(treeList, 0);
 	}
+
 }
